RadioTh_Server/RadioTh_Server/serializer.py
"""
Serializers created for models in Players_Attribute_Backend project.
"""

from rest_framework import serializers
from .models import Treatment


class TreatmentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Treatment
        fields = '__all__'

